# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: SchemaBridge API
  version: 0.1.0
servers:
  - url: http://localhost:4000
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: API status
  /schemas:
    post:
      summary: Create a schema document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSchemaRequest"
      responses:
        "200":
          description: Created schema
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemaDocument"
    get:
      summary: List schema documents
      responses:
        "200":
          description: Schemas
  /schemas/{id}:
    get:
      summary: Get schema document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Schema
        "404":
          description: Not found
    patch:
      summary: Update schema document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSchemaRequest"
      responses:
        "200":
          description: Updated schema
        "404":
          description: Not found
    delete:
      summary: Delete schema document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: cascade
          in: query
          required: false
          description: Set to "true" to delete dependent mappings and delete or update affected bindings.
          schema:
            type: string
            enum: ["true", "false"]
      responses:
        "204":
          description: Deleted
        "409":
          description: Schema is still referenced
  /mappings:
    post:
      summary: Create mapping
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMappingRequest"
      responses:
        "200":
          description: Created mapping
    get:
      summary: List mappings
      responses:
        "200":
          description: Mappings
  /mappings/{id}:
    get:
      summary: Get mapping
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Mapping
        "404":
          description: Not found
    delete:
      summary: Delete mapping
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: cascade
          in: query
          required: false
          description: Set to "true" to delete request bindings that use this mapping and clear response-only references.
          schema:
            type: string
            enum: ["true", "false"]
      responses:
        "204":
          description: Deleted
        "409":
          description: Mapping is still referenced
  /mappings/{id}/versions:
    post:
      summary: Create mapping version
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rules]
              properties:
                rules:
                  type: array
                  items:
                    $ref: "#/components/schemas/MappingRule"
      responses:
        "200":
          description: Updated mapping
  /mappings/{id}/versions/current:
    patch:
      summary: Update current mapping version
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rules]
              properties:
                rules:
                  type: array
                  items:
                    $ref: "#/components/schemas/MappingRule"
      responses:
        "200":
          description: Updated mapping without creating a new version
        "404":
          description: Mapping version not found
  /mappings/{id}/restore:
    post:
      summary: Restore mapping version
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version]
              properties:
                version:
                  type: integer
                  minimum: 1
      responses:
        "200":
          description: Updated mapping
  /transform:
    post:
      summary: Execute transformation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [input, rules]
              properties:
                input: {}
                rules:
                  type: array
                  items:
                    $ref: "#/components/schemas/MappingRule"
      responses:
        "200":
          description: Transform result
  /bindings:
    get:
      summary: List proxy bindings
      responses:
        "200":
          description: Bindings
    post:
      summary: Create proxy binding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProxyBindingRequest"
      responses:
        "200":
          description: Created binding
  /bindings/{id}:
    get:
      summary: Get proxy binding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Binding
        "404":
          description: Not found
    patch:
      summary: Update proxy binding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProxyBindingRequest"
      responses:
        "200":
          description: Updated binding
        "404":
          description: Not found
    delete:
      summary: Delete proxy binding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /apps:
    get:
      summary: List proxy apps
      responses:
        "200":
          description: Apps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProxyApp"
    post:
      summary: Create proxy app
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProxyAppRequest"
      responses:
        "200":
          description: Created app with one-time plaintext key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProxyAppWithKey"
  /apps/{id}:
    get:
      summary: Get proxy app
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: App
        "404":
          description: Not found
    patch:
      summary: Update proxy app
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProxyAppRequest"
      responses:
        "200":
          description: Updated app
        "404":
          description: Not found
    delete:
      summary: Delete proxy app
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /apps/{id}/rotate:
    post:
      summary: Rotate proxy app key
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: App with one-time fresh plaintext key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProxyAppWithKey"
        "404":
          description: Not found
  /proxy/requests:
    get:
      summary: List recent proxy request logs
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
        - name: since
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Proxy request logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProxyRequestLog"
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    CreateSchemaRequest:
      type: object
      required: [name, content]
      properties:
        name:
          type: string
        content: {}
    UpdateSchemaRequest:
      type: object
      properties:
        name:
          type: string
        content: {}
    SchemaDocument:
      type: object
      required: [id, name, content, fields, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        content: {}
        fields:
          type: array
        createdAt:
          type: string
          format: date-time
    CreateMappingRequest:
      type: object
      required: [name, sourceSchemaId, targetSchemaId, rules]
      properties:
        name:
          type: string
        sourceSchemaId:
          type: string
          format: uuid
        targetSchemaId:
          type: string
          format: uuid
        rules:
          type: array
          items:
            $ref: "#/components/schemas/MappingRule"
    MappingRule:
      type: object
      required: [id, sourcePath, targetPath]
      properties:
        id:
          type: string
        sourcePath:
          type: string
        targetPath:
          type: string
        defaultValue: {}
        transform:
          type: string
          enum: [string, number, boolean, lowercase, uppercase, iso-date]
    CreateProxyBindingRequest:
      type: object
      required: [name, method, pathPattern, upstreamBaseUrl, mappingId]
      properties:
        name:
          type: string
        method:
          type: string
          enum: [GET, POST, PUT, PATCH, DELETE, HEAD, "*"]
        pathPattern:
          type: string
        upstreamBaseUrl:
          type: string
          format: uri
        mappingId:
          type: string
          format: uuid
        responseMappingId:
          type: string
          format: uuid
          nullable: true
        forwardHeaders:
          type: array
          description: >-
            Inbound headers copied to the upstream request. Defaults to
            content-type and accept. The Authorization header used to
            authenticate to the bridge is never forwarded upstream, even if
            listed here, so proxy app keys cannot leak to upstreams.
          items:
            type: string
        validationMode:
          type: string
          enum: [off, warn, strict]
          default: off
        enabled:
          type: boolean
    UpdateProxyBindingRequest:
      allOf:
        - $ref: "#/components/schemas/CreateProxyBindingRequest"
      description: Partial proxy binding update; all fields are optional.
    ProxyApp:
      type: object
      required: [id, name, description, keyPrefix, scope, bindingIds, enabled, lastUsedAt, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        keyPrefix:
          type: string
        scope:
          type: string
          enum: [all, selected]
        bindingIds:
          type: array
          items:
            type: string
            format: uuid
        enabled:
          type: boolean
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProxyAppWithKey:
      allOf:
        - $ref: "#/components/schemas/ProxyApp"
        - type: object
          required: [key]
          properties:
            key:
              type: string
    CreateProxyAppRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 80
        description:
          type: string
          maxLength: 280
        scope:
          type: string
          enum: [all, selected]
        bindingIds:
          type: array
          items:
            type: string
            format: uuid
        enabled:
          type: boolean
    UpdateProxyAppRequest:
      allOf:
        - $ref: "#/components/schemas/CreateProxyAppRequest"
      description: Partial proxy app update; all fields are optional.
    ProxyRequestLog:
      type: object
      required: [id, bindingId, appId, method, path, statusCode, durationMs, upstreamUrl, incomingRequest, transformedRequest, responseBody, errors, createdAt]
      properties:
        id:
          type: string
          format: uuid
        bindingId:
          type: string
          format: uuid
          nullable: true
        appId:
          type: string
          format: uuid
          nullable: true
        method:
          type: string
        path:
          type: string
        statusCode:
          type: integer
        durationMs:
          type: integer
          minimum: 0
        upstreamUrl:
          type: string
          nullable: true
        incomingRequest:
          nullable: true
        transformedRequest:
          nullable: true
        responseBody:
          nullable: true
        errors:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time